Refactored Program.cs to have separate menu flows and also changed the name of the console app to Cli to avoid namespace conflicts with System.Console
