+++ TravelPlanner/TravelPlanner.Cli/Commands/TripMenuCommand.cs
@@ -1,4 +1,4 @@
-namespace TravelPlanner.Console.Commands;
+namespace TravelPlanner.Cli.Commands;
 
 public enum TripMenuCommand
 {
