+++ TravelPlanner/TravelPlanner.Cli/Parsers/BookmarkDetailMenuParser.cs
@@ -1,4 +1,4 @@
-﻿namespace TravelPlanner.ConsoleApp;
+﻿namespace TravelPlanner.Cli;
 
 public static class BookmarkDetailMenuParser
 {
