+++ TravelPlanner/TravelPlanner.Cli/Parsers/ExpenseDetailMenuParser.cs
@@ -1,4 +1,4 @@
-﻿namespace TravelPlanner.ConsoleApp;
+﻿namespace TravelPlanner.Cli;
 
 public static class ExpenseDetailMenuParser
 {
