+++ TravelPlanner/TravelPlanner.Cli/data/trips.json
