+++ TravelPlanner/TravelPlanner.slnx
@@ -1,5 +1,5 @@
 <Solution>
-  <Project Path="TravelPlanner.Console/TravelPlanner.Console.csproj" />
+  <Project Path="TravelPlanner.Cli/TravelPlanner.Cli.csproj" />
   <Project Path="TravelPlanner.Core.Tests/TravelPlanner.Core.Tests.csproj" />
   <Project Path="TravelPlanner.Core/TravelPlanner.Core.csproj" />
 </Solution>
